Move Verilog RTL synthesis tests
Keep the Verilog module facade focused by moving RTL lowering tests into rtl.rs and SynthNetlist tests into synth.rs.

Also mark the hardcoded sequential macro route fallback as a future deprecation target once searched RS/D latch placement is reliable.
